refactoring device uptime microservice
